requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
nltk>=3.6.0
spacy>=3.1.0
textblob>=0.15.0
vaderSentiment>=3.3.0
scikit-learn>=1.0.0
gensim>=4.0.0
wordcloud>=1.8.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
streamlit>=1.28.0
networkx>=2.6.0 